+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - disallow removing armor with curse of binding (fix #348)
 - add tag for disabling restock, disable restock for powah reactors (fix #349)
 - fix issue where items that stack to less than 32 would get overstacked by pick block
+- change icon for terminal settings (close #341)